Fix wrong octave
It looks like lilypond changed and the same file used in previous years
results in a shift of a few octaves at bar 24, the beginning of the second
verse. I don't recall the syntax/semantics well. I mucked with the file and
now the output looks indistinguishable from preserved output from previous years.